embeddings: one provider — OpenAI-compatible or nothing
Delete the huggingface and local providers and the EMBEDDING_PROVIDER
auto-detection chain. Which company receives customer content for
embedding is now explicit config, not a function of which API keys
happen to exist in the environment. A leftover EMBEDDING_PROVIDER
naming a removed provider fails loud at first use instead of silently
rerouting content to OpenAI.

Prod has run the OpenAI path (text-embedding-3-small @ 384 dims) since
the original embedding_service.py; this removes the unused branches.
Self-hosters point EMBEDDING_API_URL at any /v1/embeddings-compatible
endpoint; without a key, embedding features stay off.
