+++ .env.example
@@ -59,26 +59,13 @@ CORS_ORIGINS=http://localhost:3457,http://localhost:3456
 BACKEND_URL=http://localhost:3456
 
 # ── Embeddings ──────────────────────────────────────────────────
-# Provider: openai | huggingface | local | auto (default).
-# "auto" detects from available keys: OPENAI_API_KEY → openai, HF_TOKEN → huggingface, else → local.
-# With no keys set, Stash uses local sentence-transformers — no API key required.
-# EMBEDDING_PROVIDER=auto
-
-# Option 1: OpenAI (or any OpenAI-compatible API — Gemini, Cohere, Nomic, etc.)
+# One provider: an OpenAI-compatible /v1/embeddings API. OpenAI by default;
+# point EMBEDDING_API_URL at any compatible endpoint (Gemini, Cohere, Nomic, …).
+# Without a key, embedding features (semantic search, memory, map) stay off.
 # OPENAI_API_KEY=sk-proj-...
 # EMBEDDING_API_KEY=sk-...                  # optional override for embedding-only key
 # EMBEDDING_API_URL=https://api.openai.com/v1/embeddings
 # EMBEDDING_MODEL=text-embedding-3-small
-
-# Option 2: Hugging Face Inference API (any HF Hub model)
-# HF_TOKEN=hf_...
-# EMBEDDING_MODEL=BAAI/bge-small-en-v1.5          # default
-
-# Option 3: Local sentence-transformers (free, no API key, downloads model on first use)
-# EMBEDDING_PROVIDER=local
-# EMBEDDING_MODEL=all-MiniLM-L6-v2          # default
-
-# Shared
 # EMBEDDING_DIMS=384                        # vector dimensions (must match your DB schema)
 
 # ── Linear ──────────────────────────────────
